Default to WARN logging
Because running this from cron is awkward otherwise, there is no easy
way to only email on *actual* failures because Python logging goes to
stderr by default.

We could change it to log to stdout, but this is simple too; unless you
use -v you will only get output on an exception.